requirements.txt
Pillow==10.0.0
ExifRead==3.0.0
Stegano==0.11.3
